+++ CHANGELOG.md
@@ -6,7 +6,8 @@ Version 5.4-SNAPSHOT
 
 ADDED:
 
-- new BaseTreeNode class provides the same functionality as TreeNode without implementing an interface
+- new test dependency data-api in major version 3
+- new class ParentIdTreeNode without direct parent reference, but therefore instead an id and parentId 
 
 CHANGED: